Move to_features func to ml.common
