+++ src/email_spam_filter/ml/logistic_regression/functions.py
@@ -6,40 +6,14 @@ import typing
 
 import pandas as pd
 
+from email_spam_filter.ml.common import to_features
+
 if typing.TYPE_CHECKING:
     from logging import Logger
 
     from sklearn.pipeline import Pipeline
 
     from email_spam_filter.common.containers import EmailData, TagData
-
-
-def to_features(emails: list[EmailData]) -> tuple[pd.DataFrame, pd.Series[int]]:
-    """Convert a list of EmailData into feature DataFrame and label Series.
-
-    Args:
-        emails: list of EmailData objects.
-    """
-    records: list[dict[str, typing.Any]] = [
-        {
-            "id": email.id,
-            "tag": email.tag,
-            "source": email.source,
-            "subject": email.subject,
-            "body": email.body,
-            "from_addr": email.from_addr,
-            "n_links": email.n_links,
-            "n_dupe_links": email.n_dupe_links,
-            "n_rcpts": email.n_rcpts,
-            "has_attach": email.has_attach,
-            "auth_fail": email.auth_fail,
-            "unique_html_tags": email.unique_html_tags,
-        }
-        for email in emails
-    ]
-    dataframe = pd.DataFrame.from_records(records)
-    y = (dataframe["tag"] == "spam").astype(int)
-    return dataframe, y
 
 
 def training_model(model: Pipeline, emails: list[EmailData], logger: Logger) -> Pipeline:
